requirements.txt
python-dotenv>=1.0.0
httpx>=0.27.0
fastapi>=0.111.0
uvicorn>=0.29.0
